refactor: move "app" and "components" and rename "api" to "graphql"
