+++ src/components/app/app.tsx
@@ -1,5 +1,5 @@
 import { AppRouterProvider } from "src/router";
-import { AppAPIProvider } from "../api";
+import { AppAPIProvider } from "../../graphql";
 
 export const App = () => {
     return (
