+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import { App } from "./app";
+import { App } from "./components/app";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
     <React.StrictMode>
